+++ core/dbt/task/test.py
@@ -1,5 +1,5 @@
 import threading
-from typing import Dict, Any, Set, Union
+from typing import Dict, Any, Union
 
 from .compile import CompileRunner
 from .run import RunTask
@@ -21,7 +21,6 @@ from dbt.exceptions import (
 from dbt.graph import (
     ResourceTypeSelector,
     SelectionSpec,
-    UniqueId,
     parse_test_selectors,
 )
 from dbt.node_types import NodeType, RunHookType
@@ -125,17 +124,6 @@ class TestSelector(ResourceTypeSelector):
             previous_state=previous_state,
             resource_types=[NodeType.Test],
         )
-
-    def expand_selection(self, selected: Set[UniqueId]) -> Set[UniqueId]:
-        # exposures can't have tests, so this is relatively easy
-        selected_tests = set()
-        for unique_id in self.graph.select_successors(selected):
-            if unique_id in self.manifest.nodes:
-                node = self.manifest.nodes[unique_id]
-                if node.resource_type == NodeType.Test:
-                    selected_tests.add(unique_id)
-
-        return selected | selected_tests
 
 
 class TestTask(RunTask):